+++ CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
+## 0.1.4-beta (2025-04-14)
+
+
+### Features
+
+- Added more context to error message around errored integrations
+- Added docker image for integration
+- Add support for successful and failed upserts
+
+### Bug Fixes
+
+- Fixed status property mapping
+
+### Improvements
+
+- Bumped ocean version to 0.22.2
+
+
 ## 0.1.3-beta (2025-04-08)